Sunbreeze Framework
Implements the framework and moving parts for the Sunbreeze Festival
Adds a function for NPCs that allows them to always render in for the player. This is used when rendering in the decorations and fireworks.
Music will turn on in specified zones for the night, as well as fireworks begin.
The framework is implemented in this PR. Said framework allows players to trade goldfish, receive points, and redeem their points.
All NPCs involved in the sunbreeze festival will be activated for the month of August

TODO:
Goldfishing is not implemented in this PR

Style Fixes for Sunbreeze PR

Fixes styling errors

Fish eyes sanity fixc

Added missing end of line in Fish Eyes' Lua

SQL error fix

Added a mission semicolon

Style fixes + logic handling changes

Changed logic handling for always render + removed unneeded function in npcentity as it's already in lua_baseentity